+++ app/Donation.php
@@ -4,10 +4,12 @@ namespace App;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use Iatstuti\Database\Support\NullableFields;
 
 class Donation extends Model
 {
     use SoftDeletes;
+	use NullableFields;
 
     /**
      * The attributes that should be mutated to dates.
@@ -15,6 +17,11 @@ class Donation extends Model
      * @var array
      */
     protected $dates = ['deleted_at'];
+
+	protected $nullable = [
+		'purpose',
+		'reference',
+    ];
 
     function donor() {
         return $this->belongsTo('App\Donor');