+++ database/migrations/2018_03_05_193122_add_purpose_referennce_to_donation.php
@@ -1,0 +1,36 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddPurposeReferennceToDonation extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('donations', function (Blueprint $table) {
+            $table->renameColumn('origin', 'channel');
+            $table->string('purpose')->after('origin')->nullable();
+            $table->string('reference')->after('purpose')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('donations', function (Blueprint $table) {
+            $table->dropColumn('reference');
+            $table->dropColumn('purpose');
+            $table->renameColumn('channel', 'origin');
+        });
+    }
+}