+++ resources/views/donations/donations/export.blade.php
@@ -2,7 +2,9 @@
     <thead>
         <tr>
             <th>@lang('donations.date')</th>
-            <th>@lang('donations.origin')</th>
+            <th>@lang('donations.channel')</th>
+            <th>@lang('donations.purpose')</th>
+            <th>@lang('donations.reference')</th>
             <th>@lang('donations.amount')</th>
             <th>@lang('donations.exchange_amount')</th>
         </tr>
@@ -11,7 +13,9 @@
         @foreach ($donations as $donation)
             <tr>
                 <td>{{ $donation->date }}</td>
-                <td>{{ $donation->origin }}</td>
+                <td>{{ $donation->channel }}</td>
+                <td>{{ $donation->purpose }}</td>
+                <td>{{ $donation->reference }}</td>
                 <td>{{ $donation->amount }}</td>
                 <td>{{ $donation->exchange_amount }}</td>
             </tr>